requirements.txt
streamlit
PyMuPDF
pytesseract
Pillow
faiss-cpu
numpy
google-generativeai
sentence-transformers
